HSEARCH-3103 Move the sort mode to *field* sort builders only
It doesn't make sense for other sorts (score, ...).
It may make sense for distance sorts, but that's not implemented yet.
