+++ backend/elasticsearch/src/main/java/org/hibernate/search/backend/elasticsearch/search/sort/impl/AbstractElasticsearchSearchSortBuilder.java
@@ -13,23 +13,15 @@ import org.hibernate.search.engine.search.sort.spi.SearchSortBuilder;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonPrimitive;
-import org.hibernate.search.engine.search.common.SortMode;
 
 abstract class AbstractElasticsearchSearchSortBuilder implements SearchSortBuilder<ElasticsearchSearchSortBuilder>,
 		ElasticsearchSearchSortBuilder {
 
 	private static final JsonAccessor<JsonElement> ORDER_ACCESSOR = JsonAccessor.root().property( "order" );
-	private static final JsonAccessor<JsonElement> MODE_ACCESSOR = JsonAccessor.root().property( "mode" );
 	private static final JsonPrimitive ASC_KEYWORD_JSON = new JsonPrimitive( "asc" );
 	private static final JsonPrimitive DESC_KEYWORD_JSON = new JsonPrimitive( "desc" );
-	private static final JsonPrimitive SUM_KEYWORD_JSON = new JsonPrimitive( "sum" );
-	private static final JsonPrimitive AVG_KEYWORD_JSON = new JsonPrimitive( "avg" );
-	private static final JsonPrimitive MIN_KEYWORD_JSON = new JsonPrimitive( "min" );
-	private static final JsonPrimitive MAX_KEYWORD_JSON = new JsonPrimitive( "max" );
-	private static final JsonPrimitive MEDIAN_KEYWORD_JSON = new JsonPrimitive( "median" );
 
 	private SortOrder order;
-	private SortMode mode;
 
 	@Override
 	public ElasticsearchSearchSortBuilder toImplementation() {
@@ -42,11 +34,6 @@ abstract class AbstractElasticsearchSearchSortBuilder implements SearchSortBuild
 	}
 
 	@Override
-	public void mode(SortMode mode) {
-		this.mode = mode;
-	}
-
-	@Override
 	public final void buildAndAddTo(ElasticsearchSearchSortCollector collector) {
 		JsonObject innerObject = new JsonObject();
 		if ( order != null ) {
@@ -56,25 +43,6 @@ abstract class AbstractElasticsearchSearchSortBuilder implements SearchSortBuild
 					break;
 				case DESC:
 					ORDER_ACCESSOR.set( innerObject, DESC_KEYWORD_JSON );
-					break;
-			}
-		}
-		if ( mode != null ) {
-			switch ( mode ) {
-				case SUM:
-					MODE_ACCESSOR.set( innerObject, SUM_KEYWORD_JSON );
-					break;
-				case AVG:
-					MODE_ACCESSOR.set( innerObject, AVG_KEYWORD_JSON );
-					break;
-				case MIN:
-					MODE_ACCESSOR.set( innerObject, MIN_KEYWORD_JSON );
-					break;
-				case MAX:
-					MODE_ACCESSOR.set( innerObject, MAX_KEYWORD_JSON );
-					break;
-				case MEDIAN:
-					MODE_ACCESSOR.set( innerObject, MEDIAN_KEYWORD_JSON );
 					break;
 			}
 		}
