+++ backend/lucene/src/main/java/org/hibernate/search/backend/lucene/search/sort/impl/AbstractLuceneSearchSortBuilder.java
@@ -6,7 +6,6 @@
  */
 package org.hibernate.search.backend.lucene.search.sort.impl;
 
-import org.hibernate.search.engine.search.common.SortMode;
 import org.hibernate.search.engine.search.sort.dsl.SortOrder;
 import org.hibernate.search.engine.search.sort.spi.SearchSortBuilder;
 
@@ -14,8 +13,6 @@ public abstract class AbstractLuceneSearchSortBuilder implements SearchSortBuild
 		LuceneSearchSortBuilder {
 
 	protected SortOrder order;
-
-	protected SortMode multi;
 
 	@Override
 	public LuceneSearchSortBuilder toImplementation() {
@@ -25,11 +22,6 @@ public abstract class AbstractLuceneSearchSortBuilder implements SearchSortBuild
 	@Override
 	public void order(SortOrder order) {
 		this.order = order;
-	}
-
-	@Override
-	public void mode(SortMode multi) {
-		this.multi = multi;
 	}
 
 }
