+++ engine/src/main/java/org/hibernate/search/engine/search/sort/spi/SearchSortBuilder.java
@@ -6,7 +6,6 @@
  */
 package org.hibernate.search.engine.search.sort.spi;
 
-import org.hibernate.search.engine.search.common.SortMode;
 import org.hibernate.search.engine.search.sort.dsl.SortOrder;
 
 /**
@@ -19,8 +18,6 @@ import org.hibernate.search.engine.search.sort.dsl.SortOrder;
 public interface SearchSortBuilder<B> {
 
 	void order(SortOrder order);
-
-	void mode(SortMode mode);
 
 	/**
 	 * @return An implementation-specific view of this builder,
